Migrate API 29 and AndroidX

+++ app/src/main/java/com/gmail/kamille1221/tripdiary/AboutActivity.kt
@@ -5,14 +5,13 @@ import android.content.Intent
 import android.content.pm.PackageManager
 import android.net.Uri
 import android.os.Bundle
-import android.os.Environment
-import android.support.v4.app.ActivityCompat
-import android.support.v7.app.AlertDialog
-import android.support.v7.app.AppCompatActivity
 import android.text.InputType
 import android.text.TextUtils
 import android.view.MenuItem
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
+import androidx.appcompat.app.AppCompatActivity
+import androidx.core.app.ActivityCompat
 import com.google.android.gms.oss.licenses.OssLicensesMenuActivity
 import com.google.firebase.storage.FirebaseStorage
 import com.google.firebase.storage.StorageReference
@@ -35,23 +34,42 @@ class AboutActivity : AppCompatActivity() {
 		supportActionBar?.setDisplayHomeAsUpEnabled(true)
 		supportActionBar?.setHomeButtonEnabled(true)
 
-		tvVersion.text = String.format(Locale.getDefault(), "%s %s", getString(R.string.version), BuildConfig.VERSION_NAME)
+		tvVersion.text = String.format(
+			Locale.getDefault(),
+			"%s %s",
+			getString(R.string.version),
+			BuildConfig.VERSION_NAME
+		)
 
-		tvLicense.setOnClickListener { startActivity(Intent(this, OssLicensesMenuActivity::class.java)) }
+		tvLicense.setOnClickListener {
+			startActivity(
+				Intent(
+					this,
+					OssLicensesMenuActivity::class.java
+				)
+			)
+		}
 		btnBackup.setOnClickListener {
-			if (ActivityCompat.checkSelfPermission(this, android.Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+			if (ActivityCompat.checkSelfPermission(
+					this,
+					Manifest.permission.WRITE_EXTERNAL_STORAGE
+				) == PackageManager.PERMISSION_GRANTED
+			) {
 				val resource: Int = R.layout.dialog_input
 				val view = layoutInflater.inflate(resource, null)
 				val builder = AlertDialog.Builder(this)
 				builder.setView(view)
 				builder.setTitle(R.string.backup)
-				view.etInput.setHint(R.string.email)
 				view.etInput.setRawInputType(InputType.TYPE_TEXT_VARIATION_EMAIL_ADDRESS)
 				builder.setPositiveButton(R.string.confirm) { _, _ -> backupRealm(view.etInput.text.toString()) }
 				builder.setNegativeButton(R.string.cancel, null)
 				builder.create().show()
 			} else {
-				ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), SpendUtils.REQUEST_CODE_EXTERNAL_STORAGE)
+				ActivityCompat.requestPermissions(
+					this,
+					arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE),
+					SpendUtils.REQUEST_CODE_EXTERNAL_STORAGE
+				)
 				return@setOnClickListener
 			}
 		}
@@ -69,7 +87,11 @@ class AboutActivity : AppCompatActivity() {
 		return false
 	}
 
-	override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<out String>, grantResults: IntArray) {
+	override fun onRequestPermissionsResult(
+		requestCode: Int,
+		permissions: Array<out String>,
+		grantResults: IntArray
+	) {
 		if (requestCode == SpendUtils.REQUEST_CODE_EXTERNAL_STORAGE && permissions.isNotEmpty() && permissions[0] == Manifest.permission.WRITE_EXTERNAL_STORAGE && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
 			btnBackup.performClick()
 		} else {
@@ -85,18 +107,27 @@ class AboutActivity : AppCompatActivity() {
 		val realm = Realm.getDefaultInstance()
 		val mStorageRef: StorageReference = FirebaseStorage.getInstance().reference
 		try {
-			val file = File(Environment.getExternalStorageDirectory().path + "/trips.realm")
-			if (file.exists()) {
-				file.delete()
-			}
-			realm.writeCopyTo(file)
-			val photoRef: StorageReference = mStorageRef.child("backups/$email/trips.realm")
-			val uri: Uri = Uri.fromFile(file)
-			val uploadTask: UploadTask = photoRef.putFile(uri)
-			uploadTask.addOnCompleteListener {
-				if (it.isSuccessful) {
-					Toast.makeText(this, getString(R.string.toast_backup_complete), Toast.LENGTH_SHORT).show()
+			val path = applicationContext.getExternalFilesDir(null)
+			if (path != null) {
+				val file = File(path.path + "/trips.realm")
+				if (file.exists()) {
+					file.delete()
 				}
+				realm.writeCopyTo(file)
+				val photoRef: StorageReference = mStorageRef.child("backups/$email/trips.realm")
+				val uri: Uri = Uri.fromFile(file)
+				val uploadTask: UploadTask = photoRef.putFile(uri)
+				uploadTask.addOnCompleteListener {
+					if (it.isSuccessful) {
+						Toast.makeText(
+							this,
+							getString(R.string.toast_backup_complete),
+							Toast.LENGTH_SHORT
+						).show()
+					}
+				}
+			} else {
+				Toast.makeText(this, "Backup ", Toast.LENGTH_SHORT).show()
 			}
 		} catch (e: Exception) {
 			e.printStackTrace()
